+++ include/databento/publishers.hpp
@@ -87,6 +87,8 @@ enum class Venue : std::uint16_t {
   Ndex = 39,
   // Databento Equities - Consolidated
   Dbeq = 40,
+  // MIAX Sapphire
+  Sphr = 41,
 };
 
 // A source of data.
@@ -273,6 +275,8 @@ enum class Publisher : std::uint16_t {
   DbeqBasicDbeq = 59,
   // DBEQ Plus - Consolidated
   DbeqPlusDbeq = 60,
+  // OPRA - MIAX Sapphire
+  OpraPillarSphr = 61,
 };
 
 // Get a Publisher's Venue.